+++ lib/workloads-codecommit-construct/destroy.sh
@@ -5,6 +5,7 @@ export SSC_ID=$(aws iam list-service-specific-credentials --user-name $ARGOCD_US
 aws iam delete-service-specific-credential --user-name $ARGOCD_USER --service-specific-credential-id $SSC_ID
 aws iam delete-user --user-name $ARGOCD_USER
 cdk destroy workloads-codecommit-repo-eks-blueprints-workloads-cc --force
+cdk destroy workloads-codecommit-repo-eks-blueprints-workloads-cc --force
 pushd ..
 rm -rf eks-blueprints-workloads-cc
 rm -rf eks-blueprints-workloads
